[FIX] point_of_sale, pos_self_order: fix preset slots filtering
When selecting a preset in pos_Self_order the slots were not filtered.
Issues:
- Full slots were still displayed
- Slots in the past were still displayed
